telegram/requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
cassandra-driver==3.29.1
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
geomet==0.2.1.post1
idna==3.7
lz4==4.3.3
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
yarl==1.9.4